+++ CHANGELOG.md
@@ -168,6 +168,7 @@
 - Account management views : add CanEdit = false on some M2O fields (company, account, tax...) and allow to create new analytic distribution template directly.
 - Purchase Order:Allow to cancel a validated purchase order.If a stock move is already generated, cancel it.
 - PURCHASE ORDER: Updated Purchase Order Report using PDFMergeTool
+- RECONCILEGROUP : Cancel button added with unreconcile facility on reconcile-group-form.
 
 ## Bug Fixes
 - Check duplicate : Apply manual filter before duplicate check and allow to open form view.